moved data loading functions from vendors to dataloaders subpackage
