+++ nlubridge/vendors/__init__.py
@@ -1,4 +1,8 @@
+# Copyright (c) 2021 Klaus-Peter Engelbrecht, Deutsche Telekom AG
+# This software is distributed under the terms of the MIT license
+# which is available at https://opensource.org/licenses/MIT
 """Implementations of interfaces to different vendors."""
+
 import sys
 from typing import TYPE_CHECKING
